ci: auto format setup
Initial Github Actions setup, just a basic CI workflow to run the
formatter `bun run format` automatically on pushes to the main branch.

The idea is to help the contributors in not having to waste brain
cycles with code formatting, the autofix.ci bot should act as a helpful
intern that standardizes all code base to some code formatting
preferences and keep it pretty to read.