refactor: move fetch handling logic to a lib (for the tests)
SvelteKit Universal load functions can be tricky to get the types right
on vitest unit tests.
The solution for now was to separate the core
handling of a fetch logic (regardless of if it will be from server or
client or universal) into an async function and unit test that instead
of the Svelte page.