feat: display node capacity
Change types to include all the fields we are interested into, and the
rendering of the card to include node capacity displayed in BTC instead
of satoshis. The conversion is a simple one that does not take into
account potential Javascript/floating point caveats, this was left as a
TODO for a followup patch to deal with in the future.