test: write missing TODO tests
Some of the required values to be displayed were not being tested
because they were under a TODO list while the specs were not met.
Now that we have country, firstSeen, updatedAt and capacity as part of
the node card, we can perform a basic unit test to check if the expected
values are being properly rendered.